changed table name, refactored Controller
