+++ lib/Controller/MenuEntryController.php
@@ -4,10 +4,8 @@ declare(strict_types=1);
 
 namespace OCA\AppAPI\Controller;
 
-use OC\Security\CSP\ContentSecurityPolicyNonceManager;
 use OCA\AppAPI\AppInfo\Application;
 use OCA\AppAPI\Attribute\AppAPIAuth;
-use OCA\AppAPI\ProxyResponse;
 use OCA\AppAPI\Service\AppAPIService;
 use OCA\AppAPI\Service\ExAppInitialStateService;
 use OCA\AppAPI\Service\ExAppMenuEntryService;
@@ -17,7 +15,6 @@ use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http as HttpAlias;
 use OCP\AppFramework\Http\Attribute\NoAdminRequired;
 use OCP\AppFramework\Http\Attribute\NoCSRFRequired;
-use OCP\AppFramework\Http\ContentSecurityPolicy;
 use OCP\AppFramework\Http\DataDisplayResponse;
 use OCP\AppFramework\Http\DataResponse;
 use OCP\AppFramework\Http\NotFoundResponse;
@@ -25,10 +22,7 @@ use OCP\AppFramework\Http\Response;
 use OCP\AppFramework\Http\TemplateResponse;
 use OCP\AppFramework\Services\IInitialState;
 use OCP\DB\Exception;
-use OCP\Files\IMimeTypeDetector;
-use OCP\Http\Client\IResponse;
 use OCP\IRequest;
-use OCP\IURLGenerator;
 
 class MenuEntryController extends Controller {
 
@@ -38,14 +32,11 @@ class MenuEntryController extends Controller {
 	public function __construct(
 		IRequest                                  $request,
 		private IInitialState                     $initialState,
-		private IURLGenerator                     $url,
 		private ExAppMenuEntryService             $menuEntryService,
 		private ExAppInitialStateService          $initialStateService,
 		private ExAppScriptsService				  $scriptsService,
 		private ExAppStylesService				  $stylesService,
 		private AppAPIService                     $service,
-		private IMimeTypeDetector                 $mimeTypeHelper,
-		private ContentSecurityPolicyNonceManager $nonceManager,
 		private ?string                           $userId,
 	) {
 		parent::__construct(Application::APP_ID, $request);
@@ -80,120 +71,6 @@ class MenuEntryController extends Controller {
 		$this->postprocess = true;
 		$response = new TemplateResponse(Application::APP_ID, 'main');
 		return $response;
-	}
-
-	private function createProxyResponse(string $path, IResponse $response, $cache = true): ProxyResponse {
-		$content = $response->getBody();
-		$isHTML = pathinfo($path, PATHINFO_EXTENSION) === 'html';
-		if ($isHTML) {
-			$nonce = $this->nonceManager->getNonce();
-			$content = str_replace(
-				'<script',
-				"<script nonce=\"$nonce\"",
-				$content
-			);
-		}
-
-		$mime = $response->getHeader('content-type');
-		if (empty($mime)) {
-			$mime = $this->mimeTypeHelper->detectPath($path);
-			if (pathinfo($path, PATHINFO_EXTENSION) === 'wasm') {
-				$mime = 'application/wasm';
-			}
-		}
-
-		$proxyResponse = new ProxyResponse(
-			data: $content,
-			length: strlen($content),
-			mimeType: $mime,
-		);
-
-		$headersToCopy = ['Content-Disposition', 'Last-Modified', 'Etag'];
-		foreach ($headersToCopy as $element) {
-			$headerValue = $response->getHeader($element);
-			if (empty($headerValue)) {
-				$proxyResponse->addHeader($element, $headerValue);
-			}
-		}
-
-		if ($cache && !$isHTML) {
-			$proxyResponse->cacheFor(3600);
-		}
-
-		$csp = new ContentSecurityPolicy();
-		$csp->addAllowedScriptDomain($this->request->getServerHost());
-		$csp->addAllowedScriptDomain('\'unsafe-eval\'');
-		$csp->addAllowedScriptDomain('\'unsafe-inline\'');
-		$csp->addAllowedFrameDomain($this->request->getServerHost());
-		$proxyResponse->setContentSecurityPolicy($csp);
-		return $proxyResponse;
-	}
-
-	#[NoAdminRequired]
-	#[NoCSRFRequired]
-	public function ExAppProxyGet(string $appId, string $other): Response {
-		$exApp = $this->service->getExApp($appId);
-		if ($exApp === null) {
-			return new NotFoundResponse();
-		}
-		if (!$exApp->getEnabled()) {
-			return new NotFoundResponse();
-		}
-
-		$response = $this->service->aeRequestToExApp(
-			$exApp, '/' . $other, $this->userId, 'GET', request: $this->request
-		);
-		if (is_array($response)) {
-			$error_response = new Response();
-			return $error_response->setStatus(500);
-		}
-		return $this->createProxyResponse($other, $response);
-	}
-
-	#[NoAdminRequired]
-	#[NoCSRFRequired]
-	public function ExAppProxyPost(string $appId, string $other): Response {
-		$exApp = $this->service->getExApp($appId);
-		if ($exApp === null) {
-			return new NotFoundResponse();
-		}
-		if (!$exApp->getEnabled()) {
-			return new NotFoundResponse();
-		}
-
-		$response = $this->service->aeRequestToExApp(
-			$exApp, '/' . $other, $this->userId,
-			params: $this->request->getParams(),
-			request: $this->request
-		);
-		if (is_array($response)) {
-			$error_response = new Response();
-			return $error_response->setStatus(500);
-		}
-		return $this->createProxyResponse($other, $response);
-	}
-
-	#[NoAdminRequired]
-	#[NoCSRFRequired]
-	public function ExAppProxyPut(string $appId, string $other): Response {
-		$exApp = $this->service->getExApp($appId);
-		if ($exApp === null) {
-			return new NotFoundResponse();
-		}
-		if (!$exApp->getEnabled()) {
-			return new NotFoundResponse();
-		}
-
-		$response = $this->service->aeRequestToExApp(
-			$exApp, '/' . $other, $this->userId, 'PUT',
-			$this->request->getParams(),
-			request: $this->request
-		);
-		if (is_array($response)) {
-			$error_response = new Response();
-			return $error_response->setStatus(500);
-		}
-		return $this->createProxyResponse($other, $response);
 	}
 
 	/**
